Extract expandable offerings list into separate OfferingList component
