+++ src/components/CourseCard.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 
 import type { Course, CourseOffering } from '../types/types';
-import { getCourseMean, getHoursPerWeek, getMean } from '../utilities/offeringHelpers';
+import { getCourseMean } from '../utilities/offeringHelpers';
+import { OfferingList } from './OfferingList';
 import { OverallScore } from './OverallScore';
 import { RatingBar } from './RatingBar';
 
@@ -20,7 +20,6 @@ interface CourseCardProps {
 }
 
 export const CourseCard = ({ course, offerings }: CourseCardProps) => {
-  const navigate = useNavigate();
   const [expanded, setExpanded] = useState(false);
   const instructionMean = getCourseMean(offerings, 2);
   const totalResponses = offerings.reduce((sum, o) => sum + o.courseResponses, 0);
@@ -99,47 +98,7 @@ export const CourseCard = ({ course, offerings }: CourseCardProps) => {
       )}
 
       {/* Expandable Offerings List */}
-      {expanded && (
-        <div className="border-t border-gray-100 px-5 py-4">
-          <h3 className="text-sm font-semibold text-gray-900 mb-3">Offerings</h3>
-          <div className="space-y-3">
-            {offerings.map((offering) => (
-              <div
-                key={offering.id}
-                role="button"
-                tabIndex={0}
-                className="flex items-center justify-between p-3 bg-gray-50 rounded-lg cursor-pointer hover:bg-gray-100 transition-colors"
-                onClick={(e) => {
-                  e.stopPropagation();
-                  navigate(`/course/${offering.id}`);
-                }}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter' || e.key === ' ') {
-                    e.stopPropagation();
-                    navigate(`/course/${offering.id}`);
-                  }
-                }}
-              >
-                <div className="min-w-0">
-                  <div className="flex items-center gap-2 mb-1">
-                    <span className="text-xs font-medium text-purple-700">
-                      {offering.quarter} {offering.year} - Section {offering.section}
-                    </span>
-                  </div>
-                  <p className="text-sm text-gray-700">
-                    {offering.professor.firstName} {offering.professor.lastName}
-                  </p>
-                  <p className="text-xs text-gray-500">
-                    {offering.courseResponses} responses • {getHoursPerWeek(offering)}{' '}
-                    hrs/wk
-                  </p>
-                </div>
-                <OverallScore value={getMean(offering, 2)} size="sm" />
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
+      {expanded && <OfferingList offerings={offerings} />}
     </div>
   );
 };
